requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.0
plotly==5.22.0
pytest==8.0.0
scikit-learn==1.4.0
nltk==3.8.1
spacy==3.7.2
sumy==0.11.0
sentence-transformers==2.5.1
faiss-cpu==1.8.0
langchain==0.1.10
wordcloud==1.9.3
